完善部分功能

+++ application/admin/controller/Upgrade.php
@@ -22,27 +22,39 @@ class Upgrade extends Backend {
 		//更新包检测地址
 		$version = $this->version;
         $domain = "http://www.hmy3.com";
-		$url = $domain . "/upgrade/shop/{$version}";
+        $upgrade_url = "{$domain}/api/upgrade/download_upgrade/type/shop/version/" . $version;
+
+//        echo $upgrade_url;die;
         try {
 			//检测更新包
-            $result = hmCurl($url);
+            $result = hmCurl($upgrade_url);
         } catch (\Exception $e) {
             return json(["msg" => "更新包获取失败，请重试！", "code" => 400]);
         }
+
+
         $result = json_decode($result, true);
+
+        if(empty($result)){
+            return json(["msg" => "更新包获取失败，请重试！", "code" => 400]);
+        }
+
+//        var_dump($result);die;
 		
-		if($result["code"] == 400 && $this->gengxin = true){
+		if($result["code"] == 400 && $this->gengxin = true){ //循环更新完毕
 		    //更新完成后刷新配置文件
             $this->refreshFile();
 			return json(["msg" => "更新完成！请刷新页面", "code" => 200]);
 		}
 		
 		//code为400的时候代表没有更新包
-        if ($result["code"] == 400) {
+        if ($result["code"] == 400) { //没有需要更新的版本
             return json(["msg" => $result["msg"], "code" => 400]);
         }
-		$this->gengxin = true;
-		
+		$this->gengxin = true; //开始更新版本
+
+
+
 		//更新包信息
         $upgrade = $result["data"];
 
@@ -55,11 +67,20 @@ class Upgrade extends Backend {
             mkdir($dir, 0777, true);
         }
 
+        //增加更新包下载次数
+        if(!file_exists($dir . $filename)){
+            $add_url = $domain . "/api/upgrade/add_upgrade_num/id/" . $upgrade['id'];
+            hmCurl($add_url);
+        }
+
+//        echo '<pre>'; print_r($upgrade);die;
+
         /**
 		 * 下载更新包到本地并赋值文件路径变量
 		 */
         $path = file_exists($dir . $filename) ? $dir . $filename : $this->download_file($file_url, $dir, $filename);
 
+//echo $path;die;
 
         $zip = new \ZipArchive();
 		
@@ -69,15 +90,24 @@ class Upgrade extends Backend {
             try {
                 //解压文件到toPath路径下，用于覆盖差异文件
                 $zip->extractTo($toPath);
+                rmdirs($path, false); //删除更新包
             } catch (\Exception $e) {
                 return json(["msg" => "没有该目录[" . $toPath . "]的写入权限", "code" => 400]);
             }
-			
+
+
 			//文件差异覆盖完成，开始更新数据库
-			include ROOT_PATH . "/sql.php";
+            if(file_exists(ROOT_PATH . "/sql.php")){
+
+
+                include ROOT_PATH . "/sql.php";
+
+                chmod(ROOT_PATH . "/sql.php",0777);
+                unlink(ROOT_PATH . "/sql.php");
+            }
+
 			
-			chmod(ROOT_PATH . "/sql.php",0777);
-			unlink(ROOT_PATH . "/sql.php");
+
 
 			//更新后台静态文件版本
             db::name('config')->where(['name' => 'version'])->update(['value' => time()]);
@@ -90,7 +120,7 @@ class Upgrade extends Backend {
 			return $this->index(); //递归更新
 
         } else {
-            unlink($path);
+            rmdirs($path, false); //删除更新包
             return json(["msg" => "更新包解压失败，请重试！", "code" => 400]);
         }
 
@@ -145,6 +175,31 @@ class Upgrade extends Backend {
         fwrite($fp2, $img);
         fclose($fp2);
         return $filename;
+    }
+
+
+    /**
+     * 检查更新
+     */
+    public function checkUpgrade(){
+        $version = $this->version;
+
+        $upgrade_url = "http://www.hmy3.com/api/upgrade/check_upgrade/type/shop/version/" . $version;
+
+
+        try {
+            $result = json_decode(hmCurl($upgrade_url), true);
+            if(empty($result)){
+                throw new \Exception("检测失败，请点击重试");
+            }
+        }catch (\Exception $e){
+            return json(['code' => 402, 'msg' => "检测失败，请点击重试"]);
+        }
+        Cache::set('upgrade_result', $result, 3600 * 12);
+
+        return $result;
+
+
     }
